requirements.txt
cmake==3.31.4
wheel==0.45.1
openai==1.63.2
requests==2.32.3
beautifulsoup4==4.13.3
llama-cpp-python==0.2.14